+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## ?
 
 - fix: resolve structured-clone FileList cloning
+- fix: correct dbVersions cleanup in sysdb when initial upgrade transaction aborts
 
 ## 17.3.3